Log exception message instead of generic error message
It is better to log the actual error message, because a exception can have different causes. For example: In the current implementation a syntax error, in a config file, will produce 'Cannot find ${configPath}'. Because this is misleading, this commit changes the behavior to log the actual exception message (f. e. 'Unable to compile TypeScript').