fix(tools): cache SCSS files only in prod build
Just like it's done for the CSS files, SCSS files should also only
cached on production build. As things were before this patch, I would
not get a rebuild of my SCSS files (as per the progeny dependency tree)
when I updated one of them. This was problematic when working in `npm
start` or `npm run start.deving` mode, and resulted in broken UI that